requirements/base.txt
discord.py==1.3.4
python-decouple==3.3